# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/period/{id}": {
      "get": {
        "operationId": "PeriodController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsePeriodDto"
                }
              }
            }
          }
        },
        "tags": [
          "period"
        ]
      },
      "patch": {
        "operationId": "PeriodController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePeriodDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsePeriodDto"
                }
              }
            }
          }
        },
        "tags": [
          "period"
        ]
      },
      "delete": {
        "operationId": "PeriodController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsePeriodDto"
                }
              }
            }
          }
        },
        "tags": [
          "period"
        ]
      }
    },
...
    "/section": {
      "post": {
        "operationId": "SectionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSectionDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSectionDto"
                }
              }
            }
          }
        },
        "tags": [
          "section"
        ]
      }
    },
    "/section/period/{periodId}": {
      "get": {
        "operationId": "SectionController_findAll",
        "parameters": [
          {
            "name": "periodId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "departmentId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "subjectId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "teacherId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "semester",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "dayId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseSectionDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "section"
        ]
      }
    },
    "/section/period/{periodId}/sections": {
      "get": {
        "operationId": "SectionController_findSectionsByPeriod",
        "parameters": [
          {
            "name": "periodId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "departmentId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "subjectId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "teacherId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "semester",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "dayId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseSectionDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "section"
        ]
      }
    },
    "/section/{id}": {
      "get": {
        "operationId": "SectionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSectionDto"
                }
              }
            }
          }
        },
        "tags": [
          "section"
        ]
      },
      "patch": {
        "operationId": "SectionController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSectionDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSectionDto"
                }
              }
            }
          }
        },
        "tags": [
          "section"
        ]
      },
      "delete": {
        "operationId": "SectionController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSectionDto"
                }
              }
            }
          }
        },
        "tags": [
          "section"
        ]
      }
    },
    "/section/generate-report": {
      "post": {
        "operationId": "SectionController_generateReport",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateReportDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Genera un reporte Excel de secciones académicas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "section"
        ]
      }
    },
...
    "/subject": {
      "post": {
        "operationId": "SubjectController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubjectDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSubjectDto"
                }
              }
            }
          }
        },
        "tags": [
          "subject"
        ]
      },
      "get": {
        "operationId": "SubjectController_findAll",
        "parameters": [
          {
            "name": "semester",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "careerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "departmentId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseSubjectDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "subject"
        ]
      }
    },
    "/subject/{id}": {
      "get": {
        "operationId": "SubjectController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSubjectDto"
                }
              }
            }
          }
        },
        "tags": [
          "subject"
        ]
      },
      "patch": {
        "operationId": "SubjectController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubjectDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSubjectDto"
                }
              }
            }
          }
        },
        "tags": [
          "subject"
        ]
      },
      "delete": {
        "operationId": "SubjectController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSubjectDto"
                }
              }
            }
          }
        },
        "tags": [
          "subject"
        ]
      }
    },
...
  },
  "info": {
    "title": "SIGEIT-API Documentation",
    "description": "API completa para el Sistema de Gestión de Información Educativa y Tecnológica",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "IdCreateEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "Department": {
        "type": "object",
        "properties": {}
      },
...
      "Teacher": {
        "type": "object",
        "properties": {}
      },
...
      "ResponsePeriodDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "start": {
            "format": "date-time",
            "type": "string"
          },
          "end": {
            "format": "date-time",
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "stage": {
            "type": "string"
          },
          "interval": {
            "type": "number"
          },
          "duration": {
            "type": "number"
          },
          "status": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "start",
          "end",
          "startTime",
          "endTime",
          "stage",
          "interval",
          "duration",
          "status"
        ]
      },
      "UpdatePeriodDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "start": {
            "format": "date-time",
            "type": "string"
          },
          "end": {
            "format": "date-time",
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "stage": {
            "type": "string"
          },
          "interval": {
            "type": "number"
          },
          "duration": {
            "type": "number"
          },
          "copyPrevious": {
            "type": "boolean"
          },
          "status": {
            "type": "boolean"
          }
        }
      },
...
      "Period": {
        "type": "object",
        "properties": {}
      },
...
      "CreateSectionDto": {
        "type": "object",
        "properties": {
          "capacity": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "subject": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "period": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "teacher": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "status": {
            "type": "boolean"
          },
          "all": {
            "type": "boolean"
          }
        },
        "required": [
          "capacity",
          "name",
          "subject",
          "period",
          "teacher",
          "status"
        ]
      },
      "Subject": {
        "type": "object",
        "properties": {}
      },
      "ResponseSectionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "capacity": {
            "type": "number"
          },
          "inscribed": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "subjectId": {
            "type": "number"
          },
          "subject": {
            "$ref": "#/components/schemas/Subject"
          },
          "periodId": {
            "type": "number"
          },
          "period": {
            "$ref": "#/components/schemas/Period"
          },
          "teacherId": {
            "type": "number"
          },
          "teacher": {
            "$ref": "#/components/schemas/Teacher"
          },
          "status": {
            "type": "boolean"
          },
          "all": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "capacity",
          "inscribed",
          "name",
          "subjectId",
          "subject",
          "periodId",
          "period",
          "teacherId",
          "teacher",
          "status",
          "all"
        ]
      },
      "UpdateSectionDto": {
        "type": "object",
        "properties": {
          "capacity": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "subject": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "period": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "teacher": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "status": {
            "type": "boolean"
          },
          "all": {
            "type": "boolean"
          }
        }
      },
      "GenerateReportDto": {
        "type": "object",
        "properties": {
          "periodId": {
            "type": "number",
            "description": "ID del período académico",
            "example": 1
          },
          "departmentId": {
            "type": "number",
            "description": "ID del departamento",
            "example": 1
          },
          "groupBy": {
            "type": "string",
            "description": "Campo por el cual agrupar (semester o teacherName)",
            "example": "semester"
          },
          "semester": {
            "type": "number",
            "description": "ID del semestre específico",
            "example": 1
          },
          "teacherId": {
            "type": "number",
            "description": "ID del profesor específico",
            "example": 1
          },
          "status": {
            "type": "boolean",
            "description": "Estado de las secciones",
            "example": true
          }
        },
        "required": [
          "periodId",
          "departmentId"
        ]
      },
      "ReportResponseDto": {
        "type": "object",
        "properties": {
          "buffer": {
            "type": "string",
            "description": "Buffer del archivo Excel generado",
            "example": "base64-encoded-excel-file",
            "format": "binary"
          },
          "filename": {
            "type": "string",
            "description": "Nombre del archivo generado",
            "example": "reporte_secciones_2024-08-24.xlsx"
          },
          "contentType": {
            "type": "string",
            "description": "Tipo MIME del archivo",
            "example": "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
          },
          "contentLength": {
            "type": "number",
            "description": "Tamaño del archivo en bytes",
            "example": 1024
          },
          "generatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha y hora de generación del reporte",
            "example": "2024-08-24T18:30:00.000Z"
          },
          "message": {
            "type": "string",
            "description": "Mensaje de confirmación",
            "example": "Reporte generado exitosamente"
          }
        },
        "required": [
          "buffer",
          "filename",
          "contentType",
          "contentLength",
          "generatedAt",
          "message"
        ]
      },
...
      "CreateSubjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "credits": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "hours": {
            "type": "number"
          },
          "semester": {
            "type": "number"
          },
          "typeCurriculum": {
            "type": "number"
          },
          "department": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "careers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IdCreateEntity"
            }
          }
        },
        "required": [
          "name",
          "code",
          "credits",
          "hours",
          "semester",
          "typeCurriculum",
          "department",
          "careers"
        ]
      },
      "ResponseSubjectDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "credits": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "hours": {
            "type": "number"
          },
          "semester": {
            "type": "number"
          },
          "typeCurriculum": {
            "type": "number"
          },
          "departmentId": {
            "type": "number"
          },
          "department": {
            "$ref": "#/components/schemas/Department"
          },
          "status": {
            "type": "boolean"
          },
          "careerIds": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "careers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "name",
          "code",
          "credits",
          "hours",
          "semester",
          "typeCurriculum",
          "departmentId",
          "department",
          "status",
          "careerIds",
          "careers"
        ]
      },
      "UpdateSubjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "credits": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "hours": {
            "type": "number"
          },
          "semester": {
            "type": "number"
          },
          "typeCurriculum": {
            "type": "number"
          },
          "department": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "careers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IdCreateEntity"
            }
          }
        }
      },
...
    }
  }
}